requirements.txt
pyyaml
tomli_w
tomli
oyaml
nanolib
extradict
requests
pytest-dependency
